GraphCommunities.jl Documentation

Welcome to the GraphCommunities.jl documentation! This package offers tools and utilities for analyzing graph communities, providing both efficient algorithms and comprehensive visualizations.

Introduction

GraphCommunities.jl was developed to simplify the process of community detection and analysis in graphs. By integrating with Julia's rich ecosystem of libraries, GraphCommunities.jl stands as a robust tool for researchers, data analysts, and anyone interested in graph theory.

Getting Started

To install GraphCommunities.jl, run the following command:

using Pkg
Pkg.add("GraphCommunities")

GraphCommunities.community_detection — Function

community_detection(g::AbstractGraph, algo::Louvain) -> Dict{Int, Int}

Detect communities in a graph g using the Louvain algorithm, a method based on modularity optimization.

The algorithm consists of two phases that are repeated iteratively:

Local Phase: Each node is moved to the community that yields the highest modularity gain.

Aggregation Phase: A new graph is constructed where nodes represent communities from the previous phase.

These phases are repeated until the modularity ceases to increase significantly.

Arguments

g::AbstractGraph: The graph on which to detect communities.

algo::Louvain: Indicates that the Louvain algorithm should be used for community detection.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities
julia> g = karate_club_graph()
julia> community_detection(g, Louvain())

Notes

The algorithm may not return the same community structure on different runs due to its heuristic nature. However, the structures should be reasonably similar and of comparable quality.

community_detection(g::AbstractGraph, algo::KClique) -> Dict{Int, Int}

Detect communities in a graph g using the K-Clique algorithm.

The function first finds triangles (or 3-cliques) in the graph. It then constructs a k-clique graph where nodes represent triangles, and edges indicate overlap. The connected components of this k-clique graph give the communities in the original graph.

Arguments

g::AbstractGraph: The graph on which to detect communities.

algo::KClique: Indicates that the K-Clique algorithm should be used for community detection.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities
julia> g = karate_club_graph()
julia> community_detection(g, KClique())

Notes

Currently, the implementation is restricted to 3-cliques (triangles). Future versions might support other clique sizes.

community_detection(g::AbstractGraph, algo::LabelPropagation) -> Dict{Int, Int}

Detect communities in a graph g using the Label Propagation algorithm.

The algorithm works by assigning each node a unique label. Then, in each iteration, each node adopts the label that is most frequent among its neighbors. The algorithm terminates when no node changes its label.

Arguments

g::AbstractGraph: The graph on which to detect communities.

algo::LabelPropagation: Indicates that the Label Propagation algorithm should be used for community detection.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities
julia> g = karate_club_graph()
julia> community_detection(g, LabelPropagation())

Notes

The algorithm may not return the same community structure on different runs due to its heuristic nature. However, the structures should be reasonably similar and of comparable quality.

GraphCommunities.generate — Function

generate(structure::ChainedCliques)::SimpleGraph

Create a graph consisting of structure.r cliques, each of size structure.k, chained together.

structure.r represents the number of cliques.

structure.k represents the size of each clique.

Returns a SimpleGraph with the chained cliques.

generate(structure::PlantedPartition)::SimpleGraph

Generate a graph based on the planted partition model.

structure.n_communities is the number of communities.

structure.nodes_per_community denotes the number of nodes per community.

structure.pintra is the probability of an edge within a community.

structure.pinter is the probability of an edge between communities.

Returns a SimpleGraph constructed based on the planted partition model.

generate(structure::KarateClub)::SimpleGraph

Construct the famous Zachary's Karate Club graph. This graph represents the friendships between the 34 members of a karate club studied by Wayne W. Zachary in 1977.

Returns a SimpleGraph representing the Karate Club network.

GraphCommunities.draw_communities — Function

draw_communities(g::AbstractGraph, communities::Dict)

Draw the graph g with nodes colored based on their community assignments.

Arguments

g::AbstractGraph: The input graph.

communities::Dict: A dictionary mapping each vertex to its community.

Returns

A plot with nodes colored based on their community.

Note

This function will only work if each node in the graph is assigned to a community.