Graph Generators

The primary function for creating graphs to test community detection algorithms on is the generate function which implements different constructions as methods using Julia's multiple dispatch.

GraphCommunities.generate — Function

generate(structure::ChainedCliques)::SimpleGraph

Create a graph consisting of structure.r cliques, each of size structure.k, chained together.

structure.r represents the number of cliques.

structure.k represents the size of each clique.

Returns a SimpleGraph with the chained cliques.

generate(structure::PlantedPartition)::SimpleGraph

Generate a graph based on the planted partition model.

structure.n_communities is the number of communities.

structure.nodes_per_community denotes the number of nodes per community.

structure.pintra is the probability of an edge within a community.

structure.pinter is the probability of an edge between communities.

Returns a SimpleGraph constructed based on the planted partition model.

generate(structure::KarateClub)::SimpleGraph

Construct the famous Zachary's Karate Club graph. This graph represents the friendships between the 34 members of a karate club studied by Wayne W. Zachary in 1977.

Returns a SimpleGraph representing the Karate Club network.

GraphCommunities.draw_communities — Function

draw_communities(g::AbstractGraph, communities::Dict)

Draw the graph g with nodes colored based on their community assignments.

Arguments

g::AbstractGraph: The input graph.

communities::Dict: A dictionary mapping each vertex to its community.

Returns

A plot with nodes colored based on their community.

Note

This function will only work if each node in the graph is assigned to a community.