Community Detection Algorithms

The primary function that this package includes is the compute function which implements different algorithms as methods using Julia's multiple dispatch.

GraphCommunities.compute — Function

compute(algo::Louvain, g::SimpleGraph)

Detect communities in a graph g using the Louvain algorithm, a method based on modularity optimization.

The algorithm consists of two phases that are repeated iteratively:

Local Phase: Each node is moved to the community that yields the highest modularity gain.

Aggregation Phase: A new graph is constructed where nodes represent communities from the previous phase.

These phases are repeated until the modularity ceases to increase significantly.

Arguments

algo::Louvain: Indicates that the Louvain algorithm should be used for community detection.

g::SimpleGraph: The graph on which to detect communities.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities

julia> g = generate(PlantedPartition())

julia> compute(Louvain(), g)

Notes

The algorithm may not return the same community structure on different runs due to its heuristic nature. However, the structures should be reasonably similar and of comparable quality.

compute(algo::KClique, g::SimpleGraph)::Dict{Int, Int}

Detect communities in a graph g using the K-Clique algorithm.

The function first finds triangles (or 3-cliques) in the graph. It then constructs a k-clique graph where nodes represent triangles, and edges indicate overlap. The connected components of this k-clique graph give the communities in the original graph.

Arguments

algo::KClique: Indicates that the K-Clique algorithm should be used for community detection.

g::SimpleGraph: The graph on which to detect communities.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities

julia> g = generate(KarateClub())

julia> compute(KClique(), g)

Notes

Currently, the implementation is restricted to 3-cliques (triangles). Future versions might support other clique sizes.

compute(algo::LabelPropagation, g::SimpleGraph)::Dict{Int, Int}

Detect communities in a graph g using the Label Propagation algorithm.

The algorithm works by assigning each node a unique label. Then, in each iteration, each node adopts the label that is most frequent among its neighbors. The algorithm terminates when no node changes its label.

Arguments

algo::LabelPropagation: Indicates that the Label Propagation algorithm should be used for community detection.

g::SimpleGraph: The graph on which to detect communities.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities

julia> g = generate(KarateClub())

julia> compute(LabelPropagation(), g)

Notes

The algorithm may not return the same community structure on different runs due to its heuristic nature. However, the structures should be reasonably similar and of comparable quality.

compute(algo::PageRank, g::AbstractGraph)::Vector{Float64}

Compute the PageRank values of the nodes in graph g using the PageRank algorithm.

Arguments

algo::PageRank: The PageRank algorithm configuration object. This should contain properties like damping factor (d), maximum number of iterations (max_iter), and tolerance (tol).

g::AbstractGraph: The graph for which to compute the PageRank. This can be a simple graph, directed graph, or a weighted version of these.

Returns

A vector of Float64 where each entry represents the PageRank value of the corresponding node in the graph.

Details

The function uses the power iteration method to compute the PageRank values. If the graph is weighted, the weights of the edges are taken into account while calculating the rank.

The algorithm iteratively refines the PageRank values until either the maximum number of iterations is reached or the values converge within the specified tolerance.

Example

julia> g = generate(PlantedPartition())

julia> algo = PageRank(d=0.85, max_iter=100, tol=1e-6)

julia> compute(algo, g)